+++ packages/react-components/react-carousel/library/src/components/CarouselContext.types.ts
@@ -3,10 +3,11 @@ import * as React from 'react';
 
 import type { CarouselUpdateData } from '../Carousel';
 
-export type CarouselIndexChangeData = EventData<
-  'click' | 'focus',
-  React.FocusEvent | React.MouseEvent<HTMLButtonElement | HTMLAnchorElement>
-> & {
+export type CarouselIndexChangeData = (
+  | EventData<'click', React.MouseEvent<HTMLButtonElement | HTMLAnchorElement>>
+  | EventData<'focus', React.FocusEvent>
+  | EventData<'drag', PointerEvent | MouseEvent>
+) & {
   /**
    * The index to be set after event has occurred.
    */
@@ -29,7 +30,10 @@ export type CarouselContextValue = {
   subscribeForValues: (listener: (data: CarouselUpdateData) => void) => () => void;
   enableAutoplay: (autoplay: boolean) => void;
   resetAutoplay: () => void;
+  // Container with controls passed to carousel engine
   containerRef?: React.RefObject<HTMLDivElement>;
+  // Viewport without controls used for interactive functionality (draggable, pause autoplay etc.)
+  viewportRef?: React.RefObject<HTMLDivElement>;
 };
 
 /**