+++ packages/react-components/react-carousel/library/src/components/CarouselNavContainer/CarouselNavContainer.tsx
@@ -4,6 +4,7 @@ import { useCarouselNavContainer_unstable } from './useCarouselNavContainer';
 import { renderCarouselNavContainer_unstable } from './renderCarouselNavContainer';
 import { useCarouselNavContainerStyles_unstable } from './useCarouselNavContainerStyles.styles';
 import type { CarouselNavContainerProps } from './CarouselNavContainer.types';
+import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * CarouselNavContainer component - This container will provide multiple valid layout options for the underlying carousel controls
@@ -12,16 +13,7 @@ export const CarouselNavContainer: ForwardRefComponent<CarouselNavContainerProps
   const state = useCarouselNavContainer_unstable(props, ref);
 
   useCarouselNavContainerStyles_unstable(state);
-
-  /**
-   * @see https://github.com/microsoft/fluentui/blob/master/docs/react-v9/contributing/rfcs/react-components/convergence/custom-styling.md
-   *
-   * TODO: 💡 once package will become stable (PR which will be part of promoting PREVIEW package to STABLE),
-   *      - uncomment this line
-   *      - update types {@link file://./../../../../../../../packages/react-components/react-shared-contexts/library/src/CustomStyleHooksContext/CustomStyleHooksContext.ts#CustomStyleHooksContextValue}
-   *      - verify that custom global style override works for your component
-   */
-  // useCustomStyleHook_unstable('useCarouselNavContainerStyles_unstable')(state);
+  useCustomStyleHook_unstable('useCarouselNavContainerStyles_unstable')(state);
 
   return renderCarouselNavContainer_unstable(state);
 });