+++ packages/web-components/src/anchor-button/anchor-button.ts
@@ -1,14 +1,14 @@
 import { attr, FASTElement, Observable } from '@microsoft/fast-element';
 import { keyEnter } from '@microsoft/fast-web-utilities';
-import { StartEnd } from '../patterns/index.js';
 import type { StartEndOptions } from '../patterns/index.js';
+import { StartEnd } from '../patterns/index.js';
 import { applyMixins } from '../utils/apply-mixins.js';
-import { toggleState } from '../utils/element-internals.js';
+import { swapStates, toggleState } from '../utils/element-internals.js';
 import {
   AnchorAttributes,
-  type AnchorButtonAppearance,
-  type AnchorButtonShape,
-  type AnchorButtonSize,
+  AnchorButtonAppearance,
+  AnchorButtonShape,
+  AnchorButtonSize,
   type AnchorTarget,
 } from './anchor-button.options.js';
 
@@ -265,12 +265,7 @@ export class AnchorButton extends BaseAnchor {
    * @param next - the next state
    */
   public appearanceChanged(prev: AnchorButtonAppearance | undefined, next: AnchorButtonAppearance | undefined) {
-    if (prev) {
-      toggleState(this.elementInternals, `${prev}`, false);
-    }
-    if (next) {
-      toggleState(this.elementInternals, `${next}`, true);
-    }
+    swapStates(this.elementInternals, prev, next, AnchorButtonAppearance);
   }
 
   /**
@@ -289,12 +284,7 @@ export class AnchorButton extends BaseAnchor {
    * @param next - the next state
    */
   public shapeChanged(prev: AnchorButtonShape | undefined, next: AnchorButtonShape | undefined) {
-    if (prev) {
-      toggleState(this.elementInternals, `${prev}`, false);
-    }
-    if (next) {
-      toggleState(this.elementInternals, `${next}`, true);
-    }
+    swapStates(this.elementInternals, prev, next, AnchorButtonShape);
   }
 
   /**
@@ -313,12 +303,7 @@ export class AnchorButton extends BaseAnchor {
    * @param next - the next state
    */
   public sizeChanged(prev: AnchorButtonSize | undefined, next: AnchorButtonSize | undefined) {
-    if (prev) {
-      toggleState(this.elementInternals, `${prev}`, false);
-    }
-    if (next) {
-      toggleState(this.elementInternals, `${next}`, true);
-    }
+    swapStates(this.elementInternals, prev, next, AnchorButtonSize);
   }
 
   /**