+++ packages/fluent2-theme/CHANGELOG.md
@@ -1,8 +1,26 @@
 # Change Log - @fluentui/fluent2-theme
 
-This log was last generated on Thu, 17 Oct 2024 07:23:06 GMT and should not be manually modified.
+This log was last generated on Fri, 01 Nov 2024 07:23:21 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.107.111](https://github.com/microsoft/fluentui/tree/@fluentui/fluent2-theme_v8.107.111)
+
+Fri, 01 Nov 2024 07:23:21 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/fluent2-theme_v8.107.110..@fluentui/fluent2-theme_v8.107.111)
+
+### Patches
+
+- Bump @fluentui/react to v8.121.8 ([PR #33167](https://github.com/microsoft/fluentui/pull/33167) by beachball)
+
+## [8.107.110](https://github.com/microsoft/fluentui/tree/@fluentui/fluent2-theme_v8.107.110)
+
+Fri, 25 Oct 2024 07:22:06 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/fluent2-theme_v8.107.109..@fluentui/fluent2-theme_v8.107.110)
+
+### Patches
+
+- Bump @fluentui/react to v8.121.7 ([PR #33136](https://github.com/microsoft/fluentui/pull/33136) by beachball)
 
 ## [8.107.109](https://github.com/microsoft/fluentui/tree/@fluentui/fluent2-theme_v8.107.109)